chore: 🔧 separate app and server directories for front-end back-end separation
