+++ src/app/availability/[slug]/layout.tsx
