+++ src/app/availability/[slug]/page.tsx
